@@ -1,11 +1,10 @@
 name: text_style_editor
 description: TextSyleEditor is a flutter widget that help you to edit a text style dynamically.
 version: 1.0.1
-author: Mehdi Zarepour<mehdi.zarepoor@gmail.com>
 homepage: https://github.com/mehdizarepour/text_style_editor
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
